+++ src/main/java/com/yawarSoft/Repositories/AuthRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface AuthRepository extends JpaRepository<AuthEntity,Long> {
+public interface AuthRepository extends JpaRepository<AuthEntity,Integer> {
     Optional<AuthEntity> findByUsername(String username);
 }
