+++ src/main/java/com/yawarSoft/Repositories/BloodBankRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface BloodBankRepository extends JpaRepository<BloodBankEntity,Long> {
+public interface BloodBankRepository extends JpaRepository<BloodBankEntity,Integer> {
 
     @Query("SELECT e.id AS id, e.name AS name FROM BloodBankEntity e")
     List<BloodBankProjection> getBloodBankSelect();
