+++ src/main/java/com/yawarSoft/Services/Interfaces/ImageStorageService.java
@@ -6,6 +6,6 @@ import java.io.IOException;
 
 
 public interface ImageStorageService {
-    String storeImage(Long idUser, MultipartFile file) throws IOException;
+    String storeImage(Integer idUser, MultipartFile file) throws IOException;
     boolean deleteImage(String imagePath);
 }
